+++ Widget/widgetmediacontrol.h
@@ -22,7 +22,7 @@ private slots:
 
     void slot_volume_value_change(int volume);
 
-    void slot_get_media_duration(uint32_t duration);
+    void slot_get_media_duration(int64_t duration);
 
     void slot_current_video_frame_time(float time);
 
@@ -30,7 +30,7 @@ private:
     Ui::WidgetMediaControl *ui;
 
     // 视频时长
-    uint32_t mMediaDuration = 0;
+    int64_t mMediaDuration = 0;
 };
 
 #endif // WIDGETMEDIACONTROL_H
